user can register with name that is a minimum length of one character, a name with a space, a name that contains the @ symbol and the name will be formatted in the chat window, the context file has been moved to the context folder and renamed, the sizing of the forms are consistent at different viewports, if the user selects the light theme and signs out there is no more white space
