+++ src/components/BottomInputComponent.js
@@ -6,7 +6,7 @@ import { BsFillArrowUpSquareFill } from "react-icons/bs"
 import Container from "react-bootstrap/Container"
 import Navbar from "react-bootstrap/Navbar"
 import Button from "react-bootstrap/Button"
-import { ThemeContext } from "../context.js"
+import { ThemeContext } from "../context/themeContext.js"
 import useSound from "use-sound"
 import buttonSound from "../sounds/stories_sounds_boop.mp3"
 import { TbMusicOff } from "react-icons/tb"
