+++ src/components/TopNavigationBar.js
@@ -6,7 +6,7 @@ import Navbar from "react-bootstrap/Navbar"
 import Button from "react-bootstrap/Button"
 import Image from "react-bootstrap/Image"
 import "./TopNavigationBar.css"
-import { ThemeContext } from "../context.js"
+import { ThemeContext } from "../context/themeContext.js"
 import { useContext } from "react"
 import UserLogo from "../assets/sparrow-user-profile.svg"
 import useSound from "use-sound"
